+++ database/migrations/2021_10_02_091721_create_plannings_table.php
@@ -27,7 +27,7 @@ class CreatePlanningsTable extends Migration
             $table->text('combine_risk_assessment')->nullable();
             $table->text('planning_analytics')->nullable();
             $table->unsignedSmallInteger('stage')->default(0);
-            $table->integer('status')->default(0);
+            $table->unsignedSmallInteger('status')->default(0);
             $table->timestamps();
         });
     }
