+++ database/migrations/2021_10_13_122813_create_assertions_table.php
@@ -19,12 +19,12 @@ class CreateAssertionsTable extends Migration
             $table->foreign('transaction_test_id')->references('id')->on('transaction_tests')->onDelete('cascade');
             $table->unsignedInteger('company_id');
             $table->foreign('company_id')->references('id')->on('companies');
-            $table->integer('completeness')->default(0);
-            $table->integer('existence')->default(0);
-            $table->integer('accuracy')->default(0);
-            $table->integer('valuation')->default(0);
-            $table->integer('obligation_right')->default(0);
-            $table->integer('disclosure_presentation')->default(0);
+            $table->unsignedSmallInteger('completeness')->default(0);
+            $table->unsignedSmallInteger('existence')->default(0);
+            $table->unsignedSmallInteger('accuracy')->default(0);
+            $table->unsignedSmallInteger('valuation')->default(0);
+            $table->unsignedSmallInteger('obligation_right')->default(0);
+            $table->unsignedSmallInteger('disclosure_presentation')->default(0);
             $table->timestamps();
         });
     }
